requirements.txt
pymysql
flask
flask-sqlalchemy
flask-limiter
werkzeug==2.0.3
pillow
pycryptodome
gunicorn
gevent